Break out the `plugin ...` command hierarchy
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`plugin ...` hierarchy.
